requirements.txt
flask==2.3.3
Flask-RESTful==0.3.8
PyJWT==1.7.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.21
SQLAlchemy-serializer==1.4.1
six==1.16.0
flask-migrate==2.7.0
